requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
numpy
python-dotenv
pydantic
aioredis
langchain
openai
scikit-learn